Created subdirectories, added go.mod, created test checking for nodeutils
