+++ go.mod
@@ -1,0 +1,3 @@
+module example.com/m/v2
+
+go 1.15
