+++ pkg/nodeutils/node.go
@@ -1,0 +1,29 @@
+package nodeutils
+
+type Node struct {
+	prev *Node
+	next *Node
+	key  interface{}
+}
+
+type List struct {
+	head *Node
+	tail *Node
+}
+
+func (L *List) Insert(key interface{}) {
+	list := &Node{
+		next: L.head,
+		key:  key,
+	}
+	if L.head != nil {
+		L.head.prev = list
+	}
+	L.head = list
+
+	l := L.head
+	for l.next != nil {
+		l = l.next
+	}
+	L.tail = l
+}
